+++ main.go
@@ -26,7 +26,7 @@ import (
 	"golang.org/x/oauth2"
 )
 
-const version = "0.2.1"
+const version = "0.2.2"
 
 func stderr() *os.File { return os.Stderr }